SRT: update libsrt from 1.4.1 to 1.5.1

+++ trunk/src/protocol/srs_protocol_srt.cpp
@@ -221,7 +221,7 @@ srs_error_t srs_srt_nonblock(srs_srt_t srt_fd)
     return srs_success;
 }
 
-srs_error_t srs_srt_set_maxbw(srs_srt_t srt_fd, int maxbw)
+srs_error_t srs_srt_set_maxbw(srs_srt_t srt_fd, int64_t maxbw)
 {
     SET_SRT_OPT(srt_fd, SRTO_MAXBW, maxbw);
     return srs_success;
@@ -311,7 +311,7 @@ srs_error_t srs_srt_set_pbkeylen(srs_srt_t srt_fd, int pbkeylen)
     return srs_success;
 }
 
-srs_error_t srs_srt_get_maxbw(srs_srt_t srt_fd, int& maxbw)
+srs_error_t srs_srt_get_maxbw(srs_srt_t srt_fd, int64_t& maxbw)
 {
     GET_SRT_OPT(srt_fd, SRTO_MAXBW, maxbw);
     return srs_success;
